TakeDamage Method adjusted

+++ The Fallen Guardian/Assets/_Scripts/ScriptableObjects/Stats/CharacterStats.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "ScriptableObjects/CharacterStats")]
 public class CharacterStats : ScriptableObject
 {
-    public float health;
-    public float maxHealth;
+    public int health;
+    public int maxHealth;
     public float movementSpeed;
-    public float damage;
+    public int damage;
     public PlayerClass playerClass; // Possibly move this out of the characterstats and into the player
 }
